changed name of climber commands to climber2022 and then the command name
