+++ src/main/java/frc/robot/commands/Climber2022DisableCMD.java
@@ -7,36 +7,35 @@ import org.apache.logging.log4j.Logger;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class ClimberUp2022CMD extends CommandBase {
+public class Climber2022DisableCMD extends CommandBase{
     private final Climber2022 climber;
 
-    private static final Logger LOGGER = RobotLogManager.getMainLogger(Climber2022.class.getName());
+    private static final Logger LOGGER = RobotLogManager.getMainLogger(Climber2022DisableCMD.class.getName());
 
 
-    public ClimberUp2022CMD(Climber2022 climber) {
+    public Climber2022DisableCMD(Climber2022 climber) {
         this.climber = climber;
-
         addRequirements(climber);
     }
+    
 
-
-    @Override
+    @Override 
     public void initialize() {
-        LOGGER.info("Climber going up");
+        climber.disable();
+        LOGGER.info("Climber disabled");
     }
 
 
     @Override 
     public void execute() {
-        climber.up();
     }
 
     @Override 
     public void end(boolean interrupted) {}
 
 
-    @Override
+    @Override 
     public boolean isFinished() {
-        return false;
+        return true;
     }
 }
