+++ src/main/java/frc/robot/commands/Climber2022UpCMD.java
@@ -7,38 +7,36 @@ import org.apache.logging.log4j.Logger;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 
-public class ClimberEnable2022CMD extends CommandBase {
+public class Climber2022UpCMD extends CommandBase {
     private final Climber2022 climber;
 
-    private static final Logger LOGGER = RobotLogManager.getMainLogger(ClimberEnable2022CMD.class.getName());
+    private static final Logger LOGGER = RobotLogManager.getMainLogger(Climber2022.class.getName());
 
 
-
-
-    public ClimberEnable2022CMD(Climber2022 climber) {
+    public Climber2022UpCMD(Climber2022 climber) {
         this.climber = climber;
 
         addRequirements(climber);
     }
 
 
-    @Override 
+    @Override
     public void initialize() {
-        climber.enable();
-        LOGGER.info("Climber enabled");
+        LOGGER.info("Climber going up");
     }
+
 
     @Override 
-    public void execute(){
+    public void execute() {
+        climber.up();
     }
-
 
     @Override 
     public void end(boolean interrupted) {}
 
 
-    @Override 
+    @Override
     public boolean isFinished() {
-        return true;
+        return false;
     }
 }
